fixed code climate
